WIP tests, upgraded packages, bump version
