+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "ember-shell",
-  "version": "0.4.0-alpha.2",
+  "version": "0.4.1-alpha",
   "description": "An Ember addon that enables you to have a fully customizable OS-like shell, where you can run your own ember apps powered by ember-engines",
   "directories": {
     "doc": "docs",
@@ -33,12 +33,13 @@
     "ember-asset-loader": "0.2.6",
     "ember-cli": "^2.11.0-beta.1",
     "ember-cli-app-version": "^2.0.0",
+    "ember-cli-chai": "0.3.2",
     "ember-cli-dependency-checker": "^1.3.0",
     "ember-cli-htmlbars": "^1.1.1",
     "ember-cli-htmlbars-inline-precompile": "^0.3.3",
     "ember-cli-inject-live-reload": "^1.4.1",
     "ember-cli-jshint": "^2.0.1",
-    "ember-cli-mocha": "0.10.4",
+    "ember-cli-mocha": "^0.10.4",
     "ember-cli-mocha-reporter": "git+https://github.com/lolmaus/ember-cli-mocha-reporter.git#b98c2694724a6b80b9fb2244a05a46ff8ffd71bd",
     "ember-cli-preprocess-registry": "^3.0.0",
     "ember-cli-release": "^0.2.9",
@@ -61,7 +62,8 @@
     "exists-sync": "0.0.4",
     "loader.js": "^4.0.10",
     "postcss-cssnext": "^2.8.0",
-    "postcss-import": "^8.2.0"
+    "postcss-import": "^9.0.0",
+    "sinon-chai": "^2.8.0"
   },
   "dependencies": {
     "ember-cli-babel": "^5.1.7",
