--- tests/unit/system/app-test.js
@@ -1,15 +1,0 @@
-/* jshint expr:true */
-//import { expect } from 'chai';
-import { it } from 'ember-mocha';
-import { describe, /*beforeEach*/ } from 'mocha';
-//import App from 'ember-shell/system/application';
-
-describe('EmberShellApp', function() {
-
-    /*beforeEach(function() {
-      this.asApp = App.create();
-    });*/
-
-    it.skip('nothing to test here, just yet', function(){});
-
-});
